♻️ refactor(monitoring): extract codex inspection panels
Move the Codex inspection results table and logs panel into focused presentation components.

Keep the page responsible for orchestration state while reducing its rendering surface.

Validated with type-check, focused monitoring tests, lint, and Vite production build.
